オブジェクトタイプをint32_t型に変更、スコアアイテムのメモリがおかしいので原因を探る

+++ ApplicationCode/Enemies/HomingEnemy.cpp
@@ -16,7 +16,7 @@ void HomingEnemy::Initialize(const std::string modelKey, const Vector3& pos, con
 	oldPos_ = pos;
 	enemy_->SetRotation(rot);
 	enemy_->SetScale(scale_);
-	enemy_->SetObjType(Object3d::OBJType::Enemy);
+	enemy_->SetObjType((int32_t)Object3d::OBJType::Enemy);
 	enemy_->SetColType(Object3d::CollisionType::Sphere);
 	hp_ = 1;
 	shotIntervalTime_ = 60;
