+++ ApplicationCode/Enemies/StraightEnemy.cpp
@@ -14,7 +14,7 @@ void StraightEnemy::Initialize(const std::string modelKey, const Vector3& pos, c
 	oldPos_ = pos;
 	enemy_->SetRotation(rot);
 	enemy_->SetScale(scale_);
-	enemy_->SetObjType(Object3d::OBJType::Enemy);
+	enemy_->SetObjType((int32_t)Object3d::OBJType::Enemy);
 	enemy_->SetColType(Object3d::CollisionType::Sphere);
 	hp_ = 1;
 	shotIntervalTime_ = 60;
